[NCL-7921] Make Rex be able to use previous task results in future tasks
Add ability to specify in a task that you would like to also include
results from previous tasks in the payload (body)

This is done by specifying a list of `previousTaskNames` in the task
definition.

The previous task names will have to be:

- in final state
- in the same graph as the current task

to be considered valid.

If the `previousTaskNames` is specified, a list is returned as payload
if more than 1 item is being used; the results and finally the original
payload of the task.