[NCL-8001] Produce rex jar with jakarta annotations
The rex jars for now are:
- rex-api
- rex-dto

This is done using the Eclipse Transformer Maven plugin. We run the
default jar through the plugin to spit out the jar with classifier
jakarta.

You can then consume either the javax jar (default) or the jakarta jar
(with classifier jakarta) in your project.

The README is also updated on how to consume the jakarta jars in a
project's library.