[NCL-8132] Specify new roles for Rex
With the authorization initiative, we need to restrict who can access
Rex. With that in mind, we'll allow users with roles:

- pnc-users-rex-admin (mostly service accounts)
- pnc-users-admin (human PNC developers)

to interact with Rex.

This is needed as we also start using a unique service account for Rex
(NCL-8054) to trigger the task endpoints. That service account might
have special roles which not all users should be able to indirectly
invoke.