Add endpoint cancelByCorrelationId
We currently have the ability to cancel a task and all of its
dependants. This commit adds the ability to cancel all the tasks with
the same correlation id.

This reduces the number of requests a client have to do to Rex to cancel
all the tasks inside a correlation id. An example is the cancellation of
a build set in Project Newcastle, which currently just iterates through
each build to
[cancel](https://github.com/project-ncl/pnc/blob/b9669dfe0e2f0872f14db98b1d9af6b85a8c34c3/remote-build-coordinator/src/main/java/org/jboss/pnc/remotecoordinator/builder/RemoteBuildCoordinator.java#L461)
all the builds in the build set.