Rework parent to be able to dev-mode from project root.
Basically the quarkus-maven-plugin has to be present in the parent if we want to run dev mode. The goals are split between core (build) and parent(generate-code, generate-code-tests) because having 'build' goal in parent caused rest-client module to try to start in dev-mode. You cannot exclude goal defined in the parent in modules, so I had to remove it from parent and added it specifically just to core module which wants that goal.